ENTITY : basic CRUD applied

+++ src/main/java/com/ll/stopnoise/domain/noiseData/repository/NoiseDataRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface NoiseDataRepository extends JpaRepository<NoiseData, Long> {
+public interface NoiseDataRepository extends JpaRepository<NoiseData, Integer> {
 }
